Update DPK verifications
