+++ packages/server/src/extensions/devicePublicKey/verifyDevicePublicKey.test.ts
@@ -1,0 +1,1 @@
+// TODO: Implement a test for `verifyDevicePublicKey`
