+++ packages/server/src/index.ts
@@ -8,7 +8,8 @@ import { generateAuthenticationOptions } from './authentication/generateAuthenti
 import { verifyAuthenticationResponse } from './authentication/verifyAuthenticationResponse';
 import { MetadataService } from './services/metadataService';
 import { SettingsService } from './services/settingsService';
-import { verifyDpkSignature, verifyDevicePublicKey } from 'extensions/devicePubKey';
+import { verifyDpkSignature } from 'extensions/devicePublicKey/verifyDpkSignature';
+import { verifyDevicePublicKey } from 'extensions/devicePublicKey/verifyDevicePublicKey';
 import { DevicePublicKeyAuthenticatorOutput } from 'helpers/decodeAuthenticatorExtensions';
 
 export {
